requirements.txt
assemblyai
elevenlabs
openai
streamlit
whisper
sounddevice
python-dotenv